pyproject.toml
[project]
name = "assistant"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "aiohappyeyeballs>=2.6.1",
    "aiohttp>=3.12.15",
    "aiohttp-retry>=2.9.1",
    "aiosignal>=1.4.0",
    "annotated-types>=0.7.0",
    "anyio>=4.10.0",
    "async-timeout>=5.0.1",
    "attrs>=25.3.0",
    "certifi>=2025.8.3",
    "charset-normalizer>=3.4.3",
    "click>=8.1.8",
    "exceptiongroup>=1.3.0",
    "fastapi>=0.116.1",
    "frozenlist>=1.7.0",
    "h11>=0.16.0",
    "idna>=3.10",
    "multidict>=6.6.4",
    "openai>=2.3.0",
    "openai-agents>=0.1.0",
    "propcache>=0.3.2",
    "pydantic>=2.11.7",
    "pydantic-core>=2.33.2",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "requests>=2.32.5",
    "sniffio>=1.3.1",
    "starlette>=0.47.3",
    "streamlit>=1.50.0",
    "twilio>=9.7.2",
    "typing-extensions>=4.15.0",
    "typing-inspection>=0.4.1",
    "urllib3>=2.5.0",
    "uvicorn>=0.35.0",
    "websockets>=15.0.1",
    "yarl>=1.20.1",
]
